+++ lib/components/abstract/AbstractRandomRange.ts
@@ -1,0 +1,7 @@
+import AbstractRange from "./AbstractRange";
+import { RandomRangeOptionsT, RandomRangeGeneratorT } from "../types";
+export default abstract class AbstractRandomRange<OptionsT, GeneratorT> extends AbstractRange<RandomRangeOptionsT<OptionsT>, RandomRangeGeneratorT<GeneratorT>> {
+  constructor(options: RandomRangeOptionsT<OptionsT>) {
+    super(options);
+  }
+}