+++ lib/components/interfaces/IRandomNumberRangeOptions.ts
@@ -1,0 +1,6 @@
+import INumberRangeOptions from './INumberRangeOptions';
+import IRandomRangeOptions from './IRandomRangeOptions';
+
+export default interface IRandomNumberRangeOptions extends IRandomRangeOptions<INumberRangeOptions> {
+  float: boolean;
+}