+++ lib/components/interfaces/IStringRangeOptions.ts
@@ -1,0 +1,5 @@
+import IRangeOptions from "./IRangeOptions";
+
+export default interface IStringRangeOptions extends IRangeOptions<number, string> {
+  source: string | string[];
+}