+++ lib/components/ranges/MinuteRange.ts
@@ -1,0 +1,11 @@
+import { AbstractDateRange } from '../abstract';
+import { DateRangeOptionsT } from '../types';
+
+export default class MinuteRange extends AbstractDateRange {
+  constructor(options?: DateRangeOptionsT) {
+    super(
+      'm',
+      options || { start: new Date(), end: Infinity, step: 1 }
+    );
+  }
+}