+++ lib/components/types/generators/DateRangeGeneratorT.ts
@@ -1,0 +1,5 @@
+import RangeGeneratorT from "./RangeGeneratorT";
+
+type DateRangeGeneratorT = RangeGeneratorT<Date>;
+
+export default DateRangeGeneratorT;