+++ lib/components/types/generators/MergeRangeGeneratorT.ts
@@ -1,0 +1,5 @@
+import AnyRangeGeneratorT from "./AnyRangeGeneratorT";
+
+type MergeRangeGeneratorT = AnyRangeGeneratorT;
+
+export default MergeRangeGeneratorT;