+++ lib/components/types/generators/RandomRangeGeneratorT.ts
@@ -1,0 +1,5 @@
+import RangeGeneratorT from "./RangeGeneratorT";
+
+type RandomRangeGeneratorT<T> = T extends RangeGeneratorT<infer RangeT> ? T : never;
+
+export default RandomRangeGeneratorT;