+++ lib/components/types/generators/ShuffleRangeGeneratorT.ts
@@ -1,0 +1,5 @@
+import RangeGeneratorT from "./RangeGeneratorT";
+
+type ShuffleRangeGeneratorT = RangeGeneratorT<any>;
+
+export default ShuffleRangeGeneratorT;