+++ lib/components/types/generators/StringRangeGeneratorT.ts
@@ -1,0 +1,5 @@
+import RangeGeneratorT from "./RangeGeneratorT";
+
+type StringRangeGeneratorT = RangeGeneratorT<string>;
+
+export default StringRangeGeneratorT;