+++ CHANGELOG.md
@@ -1,5 +1,31 @@
+# v3.0.0
+
+- Refactoring.
+
+- Added random ranges.
+  - `RandomNumberRange`.
+  - `RandomCharRange`.
+  - `RandomStringRange`.
+  - `RandomDateRange`.
+  - `RandomColorRange`.
+- Added `ShuffleRange`.
+- Added more tests.
+- Added Wiki.
+- Added CHANGELOG.md.
+
+- Removed `merged` getter from `ZipRange`. Use `dict` instead.
+- Removed `DateRange` class. Use `MillisecondRange` instead.
+- Removed `leepYear` parameter in `DateRangeOptionsT`. Use `leapYear` instead (grammar mistake :D).
+- Removed `Range`.
+
+- Changed from `ZipRanges` to `ZipRange`.
+- Changed from `MergeRanges` to `MergeRange`.
+
+- Fix all bugs with filter function.
+
 # v2.2.0
-- For `ZipRange` use `dict` getter instead a `merged`. `merged` getter is deprecated.
+
+- For `ZipRanges` use `dict` getter instead a `merged`. `merged` getter is deprecated.
 - Added new getter `stringify`.
 - Fixed `DateRange` error
 - Added more tests